+++ lib/geodesy.dart
@@ -1,2 +1,3 @@
 export 'src/geodesy.dart';
-export 'package:latlong2/latlong.dart';
+export 'package:latlong2/latlong.dart' hide pi;
+export 'package:geodesy/src/core/core.dart';
