+++ lib/src/core/final_bearing_between_two_geo_points.dart
@@ -1,0 +1,6 @@
+import 'core.dart';
+
+/// calculate the final bearing from point l1 to point l2
+num fBearingBetweenTwoGeoPoints(LatLng l1, LatLng l2) {
+  return (bBetweenTwoGeoPoints(l2, l1) + 180) % 360;
+}
